Revert "Bug fixing in branch"
This reverts commit 0f69b91976705909c51c31ad097b10f2506f78d6.

+++ Assets/Scripts/App/BackendModelConversion/FromProtobufExtensions.cs
@@ -95,8 +95,8 @@ namespace Loom.ZombieBattleground.Data
                 (int) hero.Level,
                 (Enumerators.SetType) hero.Element,
                 hero.Skills.Select(skill => skill.FromProtobuf()).ToList(),
-                (int)hero.PrimarySkill,
-                (int)hero.SecondarySkill
+                hero.PrimarySkill,
+                hero.SecondarySkill
             );
         }
 
@@ -133,8 +133,8 @@ namespace Loom.ZombieBattleground.Data
                 (int) deck.HeroId,
                 deck.Name,
                 deck.Cards.Select(skill => skill.FromProtobuf()).ToList(),
-                (int)deck.PrimarySkill,
-                (int)deck.SecondarySkill
+                deck.PrimarySkill,
+                deck.SecondarySkill
             );
         }
